refactor project layout
move all code into 'trekapi' module
